python-backend/requirements.txt
fastapi
uvicorn
PyJWT
pydantic
passlib[bcrypt]